requirements.txt
Django~=4.2.5
matplotlib~=3.8.0
django-ckeditor~=6.7.0